# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "BodyGuard API",
    "description": "BodyGuard API"
  },
  "servers": [
    {
      "url": " https://bbb5-194-141-252-114.eu.ngrok.io",
      "description": "BodyGuard API"
    }
  ],
...
  "paths": {
...
    "/requests": {
      "post": {
        "tags": ["Requests"],
        "summary": "Create request",
        "description": "Create request",
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Request"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
    "/requests/users/{userId}": {
      "get": {
        "tags": ["Requests"],
        "summary": "Get requests by user id",
        "description": "Get requests by user id",
        "operationId": "getRequestsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "number",
              "format": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Request"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/requests/": {
      "get": {
        "tags": ["Requests"],
        "summary": "Get all requests",
        "description": "Get all requests",
        "operationId": "getAllRequests",
        "responses": {
          "200": {
            "description": "Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Request"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "400": {
        "description": "Bad Request"
      },
...
      "404": {
        "description": "Not Found"
      },
...
    },
    "schemas": {
...
      "Request": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "description": {
            "type": "string",
            "example": "Jane Doe"
          },
          "userId": {
            "type": "number",
            "example": 1
          },
          "startingAddress": {
            "type": "string",
            "example": "tuka bratmee"
          },
          "endingAddress": {
            "type": "string",
            "example": "1234567890"
          },
          "isGuard": {
            "type": "boolean",
            "example": false
          },
          "isDriver": {
            "type": "boolean",
            "example": false
          },
          "hasGun": {
            "type": "boolean",
            "example": false
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "startingTime": {
            "type": "string",
            "example": "2020-10-10T10:00:00.000Z"
          }
        }
      }
    }
  }
}